+++ tests/Controller/ShowTest.php
@@ -2,9 +2,14 @@
 
 namespace TomHart\Restful\Tests\Controller;
 
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Foundation\Testing\TestResponse;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Mockery\Mock;
+use Mockery\MockInterface;
 use Symfony\Component\HttpFoundation\Response as SymResponse;
+use TomHart\Restful\AbstractRestfulController;
 use TomHart\Restful\Tests\Classes\Models\ModelTest;
 use TomHart\Restful\Tests\TestCase;
 
@@ -21,11 +26,17 @@ class ShowTest extends TestCase
         $model->save();
 
         /** @var TestResponse $response1 */
-        $response1 = $this->get(route('model-tests.show', [
-            'model_test' => $model->id
-        ]), [
-            'Accept' => 'application/json'
-        ]);
+        $response1 = $this->get(
+            route(
+                'model-tests.show',
+                [
+                    'model_test' => $model->id
+                ]
+            ),
+            [
+                'Accept' => 'application/json'
+            ]
+        );
 
         /** @var JsonResponse $response */
         $response = $response1->baseResponse;
@@ -46,11 +57,41 @@ class ShowTest extends TestCase
         $model->name = 'Test 1';
         $model->save();
 
-        $response = $this->get(route('model-tests.show', [
-            'model_test' => $model->id
-        ]));
+        $response = $this->get(
+            route(
+                'model-tests.show',
+                [
+                    'model_test' => $model->id
+                ]
+            )
+        );
 
         $this->responseIsHtml($response);
         $this->assertEquals(SymResponse::HTTP_OK, $response->baseResponse->getStatusCode());
+    }
+
+    /**
+     * Test show supports string ids.
+     */
+    public function testShowSupportsStringIds(): void
+    {
+        $this->expectException(ModelNotFoundException::class);
+        $this->expectExceptionMessage('No query results for model [TomHart\Restful\Tests\Classes\Models\ModelTest] abc');
+        /** @var Request|MockInterface $mockRequest */
+        $mockRequest = $this->mock(Request::class);
+
+        $mockRequest->shouldReceive('wantsJson')->andReturn(true);
+
+        /** @var Mock|AbstractRestfulController $mockController */
+        $mockController = $this
+            ->mock(AbstractRestfulController::class)
+            ->makePartial()
+            ->shouldAllowMockingProtectedMethods();
+
+        $mockController
+            ->shouldReceive('getModelClass')
+            ->andReturn(ModelTest::class);
+
+        $mockController->show($mockRequest, 'abc');
     }
 }