+++ CHANGELOG.md
@@ -10,7 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support for ?limit
-- `findModel` supports all types rather than just int now. 
+- `findModel` supports all types rather than just int now.
+
+### Changed
+- OPTIONS with ID requested via query string now when using `Builder`
 
 ## [1.5.0] - 2020-01-14
 ### Added